♻️ refactor : 댓글 및 게시글 위치 수정
