+++ src/main/java/com/blog/workspace/adapter/in/web/CommentController.java
@@ -1,6 +1,6 @@
 package com.blog.workspace.adapter.in.web;
 
-import com.blog.workspace.application.in.CommentUseCase;
+import com.blog.workspace.application.in.comment.CommentUseCase;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
