+++ src/main/java/com/blog/workspace/application/in/post/PostUseCase.java
@@ -1,4 +1,4 @@
-package com.blog.workspace.application.in;
+package com.blog.workspace.application.in.post;
 
 public interface PostUseCase {
